requirement.txt
boto3
mysql.connector
flask